Main/requirements.txt
streamlit
google-generativeai
pandas
json